pyproject.toml
[tool.poetry]
name = "questionnaire-webtool"
version = "0.0.1"
description = "A tool for the setup of an automated questionnaire as a webtool"
authors = ["JannisZeller"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1"
flask = "^3"
flask-sqlalchemy = "^3.1.1"
werkzeug = "^3"
flask-wtf = "^1.2.1"
requests = "^2.31.0"
pyjwt = "^2"
email_validator = "^2.1"
scikit-learn = "^1.3.1"
pandas = "^2"
transformers = "^4"
Pillow = "^10"
seaborn = "^0.13"
torchvision = {version = "^0.17.1+cpu", source = "pytorch"}
ipykernel = "^6"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
